Separate RangeManager into trait and implementation
